requirements.txt
streamlit==1.39.0
uuid==1.30
python-dateutil==2.8.2
streamlit-extras==0.5.0
openai==1.35.3
streamlit_lottie==0.0.5
python-dotenv